+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- You can now configure the add-on to use a local instance of the Skybrush
+  Studio server to run transition calculations and `.skyc` exports locally.
+
 ## [2.1.0] - 2022-06-17
 
 ### Added